moves postImage to utils/helperFunctions
